NXT-101 Rename all occurrences of GuidanceListPage to ProductListPage and GuidanceListNav to ProductListNav and move directories
