+++ web/src/components/ProductListPage/ProductListGetServerSideProps/ProductListGetServerSideProps.test.ts
@@ -7,9 +7,9 @@ import {
 import { logger } from "@/logger";
 import sampleData from "@/mockData/search/guidance-published.json";
 
-import { GuidanceListPageProps } from "../GuidanceListPageProps";
+import { ProductListPageProps } from "../ProductListPageProps";
 
-import { getGetServerSidePropsFunc } from "./GuidanceListGetServerSideProps";
+import { getGetServerSidePropsFunc } from "./ProductListGetServerSideProps";
 
 import type { GetServerSidePropsContext, Redirect } from "next";
 import type { ParsedUrlQuery } from "querystring";
@@ -24,6 +24,7 @@ describe("getGetServerSidePropsFunc", () => {
 		defaultSortOrder: SortOrder.dateAscending,
 		gstPreFilter: "Published",
 		dateFilterLabel: "Last updated date",
+		index: "guidance",
 	});
 
 	describe("Redirects", () => {
@@ -85,7 +86,7 @@ describe("getGetServerSidePropsFunc", () => {
 	});
 
 	describe("Success", () => {
-		let result: { props: GuidanceListPageProps };
+		let result: { props: ProductListPageProps };
 		const resolvedUrl =
 			"/guidance/published?q=test&ndt=Guidance&from=2020-07-28&to=2021-06-04";
 		const setHeader = jest.fn();
@@ -101,7 +102,7 @@ describe("getGetServerSidePropsFunc", () => {
 				},
 				resolvedUrl,
 			} as GetServerSidePropsContext)) as {
-				props: GuidanceListPageProps;
+				props: ProductListPageProps;
 			};
 		});
