+++ web/src/pages/guidance/topic-selection.page.tsx
@@ -1,11 +1,11 @@
 import { Breadcrumb } from "@nice-digital/nds-breadcrumbs";
 import { SortOrder, Document } from "@nice-digital/search-client";
 
-import { GuidanceListNav } from "@/components/GuidanceListPage/GuidanceListNav/GuidanceListNav";
+import { GuidanceListNav } from "@/components/ProductListNav/GuidanceListNav";
 import {
-	getGuidanceListPage,
+	getProductListPage,
 	getGetServerSidePropsFunc,
-} from "@/components/GuidanceListPage/GuidanceListPage";
+} from "@/components/ProductListPage/ProductListPage";
 
 const defaultSortOrder = SortOrder.titleAscending;
 
@@ -46,7 +46,7 @@ const tableBodyRender = (documents: Document[]) => (
 	</>
 );
 
-export default getGuidanceListPage({
+export default getProductListPage({
 	metaDescription: "Find out what guidance is being considered for development",
 	listNavType: GuidanceListNav,
 	breadcrumbTrail: [
